requirements.txt
requests
asgiref==3.4.1
Django==3.2.8
powerline-status==2.7
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
django-cors-headers==3.10.1